Added typecasting to resolve compilation problem on 32-bit systems, where sizeof(void*) != sizeof(int64)

+++ src/runtime/eval/debugger/debugger_proxy.cpp
@@ -115,7 +115,7 @@ void DebuggerProxy::switchThreadMode(ThreadMode mode,
     m_thread = 0;
     notify();
   } else {
-    m_thread = Process::GetThreadId();
+    m_thread = (int64)Process::GetThreadId();
   }
   if (mode == Normal) {
     m_jump.reset();
@@ -556,7 +556,7 @@ const char *DebuggerProxy::getThreadType() const {
 
 DThreadInfoPtr DebuggerProxy::createThreadInfo(const std::string &desc) {
   DThreadInfoPtr info(new DThreadInfo());
-  info->m_id = Process::GetThreadId();
+  info->m_id = (int64)Process::GetThreadId();
   info->m_desc = desc;
   Transport *transport = g_context->getTransport();
   if (transport) {
